会計明細入力での小数点切り捨て追加に向け、小数点第2位で丸める処理を分離
