+++ lib/ui/settle_accounts/settle_accounts_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:tatetsu/model/entity/creditor.dart';
+import 'package:tatetsu/model/core/double_ext.dart';
 import 'package:tatetsu/model/entity/participant.dart';
 import 'package:tatetsu/model/entity/payment.dart';
 import 'package:tatetsu/model/entity/settlement.dart';
